+++ gulpfile.ts
@@ -14,6 +14,7 @@ const projectOptions: buildTools.Project = {
   typescript: {
     tsconfigJson: ["tsconfig.json"],
     compilerOptions: {
+      declaration: true,
       module: "commonjs",
       typeRoots: [
         "src/custom-typings",