+++ src/test/index.spec.ts
@@ -1,5 +1,3 @@
-import {assert} from "chai";
-
 describe("WebBuildTools", function () {
   it("should compile", function () {
   });